in adc.c: - correction of function pointer return type

+++ mj8x8/Core/Inc/adc/adc.h
@@ -4,7 +4,7 @@
 typedef struct adc_t	// struct describing the event handler functionality
 {
 	void (*ReadChannels)(void);  // reads channels from ADC ISR
-	uint8_t (*GetVal)(const uint8_t);  // returns channel reading at index
+	uint16_t (*GetVal)(const uint8_t);  // returns channel reading at index
 } adc_t;
 
 #endif /* CORE_INC_ADC_ADC_H_ */
